PN-203: Group the matched phenotypes under the broader system branches, e.g. Abnormality of the head, Abnormality of the neck
- Separated genotype similarity to own component
