+++ pom.xml
@@ -70,6 +70,8 @@
     <module>similarity-data-api</module>
     <module>similarity-data-impl</module>
     <module>similarity-data-rest</module>
+    <module>similarity-genotype-view</module>
+    <module>similarity-phenotype-view</module>
     <module>similarity-search</module>
     <module>communication</module>
     <module>ui</module>
@@ -296,6 +298,48 @@
               <differenceType>7012</differenceType>
               <method>org.phenotips.data.permissions.AccessLevel getAccess()</method>
               <justification>Added to determine what type of access does the user have to this patient profile.</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/data/similarity/FeatureClusterView</className>
+              <differenceType>7012</differenceType>
+              <method>java.lang.String getId()</method>
+              <justification>Added to determine the root/ancestor id of the cluster.</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/data/similarity/FeatureClusterView</className>
+              <differenceType>7012</differenceType>
+              <method>java.lang.String getName()</method>
+              <justification>Added to determine the root/ancestor name of the cluster.</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/data/similarity/FeatureClusterView</className>
+              <differenceType>7002</differenceType>
+              <method>double getScore()</method>
+              <justification>Relocated to PatientSimilarityView.</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/data/similarity/FeatureClusterView</className>
+              <differenceType>4001</differenceType>
+              <to>org/phenotips/data/Feature</to>
+              <justification>Removed org.phenotips.data.Feature from the set of implemented interfaces</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/data/similarity/FeatureClusterView</className>
+              <differenceType>4001</differenceType>
+              <to>org/phenotips/data/VocabularyProperty</to>
+              <justification>Removed org.phenotips.data.VocabularyProperty from the set of implemented interfaces</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/matchingnotification/match/PhenotypesMap</className>
+              <differenceType>8001</differenceType>
+              <justification>Relocating PhenotypesMap to new phenotype-view component.</justification>
+            </difference>
+            <difference>
+              <className>org/phenotips/matchingnotification/match/PatientInMatch</className>
+              <differenceType>7006</differenceType>
+              <method>org.phenotips.matchingnotification.match.PhenotypesMap getPhenotypes()</method>
+              <to>org.phenotips.data.similarity.phenotype.PhenotypesMap</to>
+              <justification>Relocating PhenotypesMap to new phenotype-view component.</justification>
             </difference>
           </ignored>
           <excludes>
