+++ similarity-data-impl/src/main/resources/META-INF/components.txt
@@ -1,6 +1,4 @@
-org.phenotips.data.similarity.internal.DefaultPatientGenotypeManager
 org.phenotips.data.similarity.internal.DefaultPatientSimilarityViewFactory
-org.phenotips.data.similarity.internal.ExomiserExomeManager
 org.phenotips.data.similarity.internal.RestrictedPatientSimilarityViewFactory
 org.phenotips.data.similarity.internal.DefaultMatchedPatientClusterView
 org.phenotips.data.similarity.script.ExomiserViewScriptService
