+++ similarity-genotype-view/src/main/resources/META-INF/components.txt
@@ -1,0 +1,2 @@
+org.phenotips.data.similarity.internal.ExomiserExomeManager
+org.phenotips.data.similarity.genotype.DefaultPatientGenotypeManager
